requirements.txt
Django==2.2
django-multiupload==0.5.2
django-upload-validator==1.1.5
django-widget-tweaks==1.4.5
Pillow==7.0.0
python-magic==0.4.15
pytz==2019.3
sqlparse==0.3.0
